requirements.txt
# 基础依赖
colorlog>=6.7
psycopg2-binary>=2.9
pandas>=1.5
akshare>=1.10
scikit-learn>=1.7.2
joblib>=1.5.2
tenacity>=9.1.2
PyYAML>=6.0

# 机器学习框架
torch>=2.0.0
torchvision>=0.15.0

# Web框架（可选，用于API）
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 数据可视化（可选）
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# 数据库和缓存
redis>=5.0.0
sqlalchemy>=2.0.0

# 云服务（阿里云）
alibabacloud_pai_dlc20201203>=1.0.0
alibabacloud_oss20190517>=2.1.0
alibabacloud_pai_eas20210701>=1.0.0

# 开发和测试
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0